@@ -1,15 +1,16 @@
 {
   "name": "lifescope-api",
-  "version": "3.5.3",
-  "description": "lifescope.io",
-  "author": "(https://lifescope.io)",
+  "version": "4.0.0",
+  "description": "Search and Explore the internet of you. lifescope.io",
+  "author": "LifeScope Labs (https://lifescope.io)",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/lifescopelabs/lifescope-api.git"
   },
   "main": "/server.js",
   "engines": {
-    "node": "^10.*",
+    "node": "^12.*",
     "yarn": "^1.5.1"
   },
   "files": [
