@@ -1,9 +1,9 @@
 'use strict';
 
-const Tokens = require('csrf');
-const _ = require('lodash');
-const config = require('config');
-const httpErrors = require('http-errors');
+import Tokens from 'csrf';
+import _ from 'lodash';
+import config from 'config';
+import httpErrors from 'http-errors';
 
 
 let options = _.pick(config.sessions.csrf, ['saltLength', 'secretLength']);
@@ -42,6 +42,8 @@ async function validate(req, res, next) {
 				(req.headers['x-xsrf-token']);
 
 			if (!tokens.verify(req.session.csrf_secret, csrftoken)) {
+				console.log(req.session);
+				console.log(req.body);
 				throw httpErrors(403, 'Invalid CSRF token.');
 			}
 
@@ -61,7 +63,7 @@ async function validate(req, res, next) {
 }
 
 
-module.exports = {
+export default {
 	create: create,
 	tokens: tokens,
 	validate: validate
