@@ -1,8 +1,8 @@
 'use strict';
 
-const _ = require('lodash');
-const httpErrors = require('http-errors');
-const multer = require('multer');
+import _ from 'lodash';
+import httpErrors from 'http-errors';
+import multer from 'multer';
 
 
 let upload = multer({
@@ -10,7 +10,7 @@ let upload = multer({
 });
 
 
-module.exports = function(req, res, next) {
+export default function(req, res, next) {
 	upload.single('spec')(req, res, function() {
 		if (_.has(req, 'file.buffer')) {
 			next();
@@ -19,4 +19,4 @@ module.exports = function(req, res, next) {
 			next(httpErrors(400, 'You must upload exactly one file.'));
 		}
 	});
-};
+}
