@@ -1,6 +1,6 @@
 'use strict';
 
-import Mandrill from 'mandrill-api/mandrill';
+import Mandrill from 'mandrill-api/mandrill.js';
 import _ from 'lodash';
 import config from 'config';
 
