@@ -2,7 +2,7 @@
 
 import _ from 'lodash';
 
-import createInitialSearches from '../util/create-initial-searches';
+import createInitialSearches from '../util/create-initial-searches.js';
 
 
 export default async function(req, res, next) {
