@@ -2,7 +2,7 @@
 
 'use strict';
 
-import { UserTC } from '../../schema/models/users';
+import { UserTC } from '../../schema/models/users.js';
 
 
 let re_token = /^Key ([a-fA-F0-9]{32})$/;
