@@ -1,6 +1,6 @@
 'use strict';
 
-import uuid from '../util/uuid';
+import uuid from '../util/uuid.js';
 
 
 export default function(req, res, next) {
