@@ -2,7 +2,7 @@ import _ from 'lodash';
 import httpErrors from 'http-errors';
 
 
-export default function(Resolver, scopes, resolvers) {
+export default function(scopes, resolvers) {
 	const secureResolvers = {};
 	let scopeArray = Array.isArray(scopes) ? scopes : [scopes];
 
