@@ -1,7 +1,7 @@
 import httpErrors from 'http-errors';
 
 
-export default function(Resolver, resolvers) {
+export default function(resolvers) {
 	const secureResolvers = {};
 
 	Object.keys(resolvers).forEach(k => {
