@@ -2,9 +2,9 @@
 
 import moment from 'moment';
 
-import { ConnectionTC } from '../../schema/models/connections';
-import { OAuthTokenTC } from "../../schema/models/oauth-tokens";
-import { UserTC } from '../../schema/models/users';
+import { ConnectionTC } from '../../schema/models/connections.js';
+import { OAuthTokenTC } from '../../schema/models/oauth-tokens.js';
+import { UserTC } from '../../schema/models/users.js';
 
 
 let re_token = /^Bearer ([a-fA-F0-9]{64})$/;
