@@ -1,7 +1,7 @@
 import moment from 'moment';
 
-import { SessionTC } from '../../schema/models/sessions';
-import { UserTC } from '../../schema/models/users';
+import { SessionTC } from '../../schema/models/sessions.js';
+import { UserTC } from '../../schema/models/users.js';
 
 
 export default async function(cookie) {
