@@ -6,9 +6,9 @@ import config from 'config';
 import moment from 'moment';
 import useragent from 'useragent'
 
-import uuid from './util/uuid';
+import uuid from './util/uuid.js';
 
-import { SessionTC } from '../schema/models/sessions';
+import { SessionTC } from '../schema/models/sessions.js';
 
 
 let options = _.pick(config.csrf, ['saltLength', 'secretLength']);
