@@ -4,11 +4,11 @@ import crypto from 'crypto';
 import _ from 'lodash';
 import moment from 'moment';
 
-import { readFile } from './fs';
-import uuid from './uuid';
-import sortDictionary from './sort-dictionary';
-import { SearchTC } from "../../schema/models/searches";
-import { UserTC } from '../../schema/models/users';
+import { readFile } from './fs.js';
+import uuid from './uuid.js';
+import sortDictionary from './sort-dictionary.js';
+import { SearchTC } from '../../schema/models/searches.js';
+import { UserTC } from '../../schema/models/users.js';
 
 
 let initialSearches = readFile('fixtures/searches/initial_searches.json');
