@@ -3,10 +3,10 @@
 import _ from 'lodash';
 import moment from 'moment';
 
-import uuid from './uuid';
+import uuid from './uuid.js';
 
-import { ContactTC} from "../../schema/models/contacts";
-import {PeopleTC} from '../../schema/models/people';
+import { ContactTC} from '../../schema/models/contacts.js';
+import {PeopleTC} from '../../schema/models/people.js';
 
 let providerMap = {
 	dropbox: {
