@@ -3,14 +3,14 @@
 import _ from 'lodash';
 import httpErrors from 'http-errors';
 
-import { ContactTC } from '../../schema/models/contacts';
-import { ConnectionTC } from '../../schema/models/connections';
-import { ContentTC } from '../../schema/models/content';
-import { EventTC } from '../../schema/models/events';
-import { LocationTC } from '../../schema/models/locations';
-import { OAuthTokenTC } from '../../schema/models/oauth-tokens';
-import { PeopleTC } from '../../schema/models/people';
-import { ThingTC } from '../../schema/models/things';
+import { ContactTC } from '../../schema/models/contacts.js';
+import { ConnectionTC } from '../../schema/models/connections.js';
+import { ContentTC } from '../../schema/models/content.js';
+import { EventTC } from '../../schema/models/events.js';
+import { LocationTC } from '../../schema/models/locations.js';
+import { OAuthTokenTC } from '../../schema/models/oauth-tokens.js';
+import { PeopleTC } from '../../schema/models/people.js';
+import { ThingTC } from '../../schema/models/things.js';
 
 
 let typeComposers = [
