@@ -1,6 +1,6 @@
 import bson from 'bson';
 import mongoose from 'mongoose';
-import nodeUUID from 'uuid/v4';
+import nodeUUID from 'uuid/v4.js';
 import uuidParse from 'uuid-parse';
 
 export default function(value) {
