@@ -2,7 +2,7 @@ import Validator from 'ajv';
 import _ from 'lodash';
 import refParser from 'json-schema-ref-parser';
 
-import { find } from './util/fs';
+import { find } from './util/fs.js';
 
 
 export const loadValidator = function(schemas) {
