@@ -3,7 +3,7 @@
 import express from 'express';
 import httpErrors from 'http-errors';
 
-import { OAuthTokenTC } from '../../../schema/models/oauth-tokens';
+import { OAuthTokenTC } from '../../../schema/models/oauth-tokens.js';
 
 let router = express.Router();
 let accessToken = router.route('/');
