@@ -5,9 +5,9 @@ import crypto from 'crypto';
 
 import express from 'express';
 
-import { ConnectionTC } from '../../../schema/models/connections';
-import { ProviderTC } from '../../../schema/models/providers';
-import { UserTC } from '../../../schema/models/users';
+import { ConnectionTC } from '../../../schema/models/connections.js';
+import { ProviderTC } from '../../../schema/models/providers.js';
+import { UserTC } from '../../../schema/models/users.js';
 
 let router = express.Router();
 let deauth = router.route('/:providerId');
