@@ -6,12 +6,12 @@ import crypto from 'crypto';
 import config from 'config';
 import express from 'express';
 
-import uuid from '../../../lib/util/uuid';
+import uuid from '../../../lib/util/uuid.js';
 
-import { ConnectionTC } from '../../../schema/models/connections';
-import { DataRemovalConfirmationsTC } from '../../../schema/models/data-removal-confirmations';
-import { ProviderTC } from '../../../schema/models/providers';
-import { UserTC } from '../../../schema/models/users';
+import { ConnectionTC } from '../../../schema/models/connections.js';
+import { DataRemovalConfirmationsTC } from '../../../schema/models/data-removal-confirmations.js';
+import { ProviderTC } from '../../../schema/models/providers.js';
+import { UserTC } from '../../../schema/models/users.js';
 
 let router = express.Router();
 let deauth = router.route('/:providerId');
