@@ -2,9 +2,9 @@
 
 import express from 'express';
 
-import accessToken from './access-token';
-import deauthorizeConnection from './deauthorize-connection';
-import deleteConnection from './delete-connection';
+import accessToken from './access-token.js';
+import deauthorizeConnection from './deauthorize-connection.js';
+import deleteConnection from './delete-connection.js';
 
 
 let router = express.Router();
