@@ -6,15 +6,15 @@ import express from 'express';
 import httpErrors from 'http-errors';
 import moment from 'moment';
 
-import uuid from '../../util/uuid';
-import { Create as CreateSession } from '../../sessions';
-import createSelfContact from '../../util/create-self-contact'
+import uuid from '../../util/uuid.js';
+import { Create as CreateSession } from '../../sessions.js';
+import createSelfContact from '../../util/create-self-contact.js'
 
-import { AssociationSessionTC } from '../../../schema/models/association-sessions';
-import { ConnectionTC } from '../../../schema/models/connections';
-import { ProviderTC } from '../../../schema/models/providers';
-import { SessionTC } from '../../../schema/models/sessions';
-import { UserTC } from '../../../schema/models/users';
+import { AssociationSessionTC } from '../../../schema/models/association-sessions.js';
+import { ConnectionTC } from '../../../schema/models/connections.js';
+import { ProviderTC } from '../../../schema/models/providers.js';
+import { SessionTC } from '../../../schema/models/sessions.js';
+import { UserTC } from '../../../schema/models/users.js';
 
 
 let router = express.Router();
