@@ -2,7 +2,7 @@
 
 import express from 'express';
 
-import csrf from '../middleware/csrf';
+import csrf from '../middleware/csrf.js';
 
 let router = express.Router();
 let csrfRoute = router.route('/');
@@ -10,7 +10,7 @@ let csrfRoute = router.route('/');
 
 async function csrfHandler(req, res) {
 	res.json({
-		csrf_token: res.context.csrf_token
+		csrf_token: res.context ? res.context.csrf_token : ''
 	});
 }
 
