@@ -3,7 +3,7 @@
 import express from 'express';
 
 import EmailClient from '../../extensions/mandrill/email-client.js';
-import { EmailUpdateRequestTC } from '../../../schema/models/email-update-request';
+import { EmailUpdateRequestTC } from '../../../schema/models/email-update-request.js';
 import { UserTC } from '../../../schema/models/users.js';
 
 let router = express.Router();
