@@ -2,7 +2,7 @@
 
 import express from 'express';
 
-import uploadFile from './upload-file';
+import uploadFile from './upload-file.js';
 
 
 let router = express.Router();
