@@ -4,12 +4,12 @@ import express from 'express';
 import httpErrors from 'http-errors';
 import mongoose from 'mongoose';
 
-import { ConnectionTC } from '../../../schema/models/connections';
-import { EventTC } from '../../../schema/models/events';
-import csrf from '../../middleware/csrf';
-import multipartParser from '../../middleware/multipart-parser';
-import uuid from '../../util/uuid';
-import { ProviderTC } from '../../../schema/models/providers';
+import { ConnectionTC } from '../../../schema/models/connections.js';
+import { EventTC } from '../../../schema/models/events.js';
+import csrf from '../../middleware/csrf.js';
+import multipartParser from '../../middleware/multipart-parser.js';
+import uuid from '../../util/uuid.js';
+import { ProviderTC } from '../../../schema/models/providers.js';
 
 let router = express.Router();
 
