@@ -1,14 +1,14 @@
 import express from 'express';
 
-import auth from './auth';
-import connections from './connections';
-import csrf from './csrf';
-import email from './email';
-import financials from './financials';
-import locations from './locations';
-import login from './login';
-import logout from './logout';
-import signup from './signup';
+import auth from './auth/index.js';
+import connections from './connections/index.js';
+import csrf from './csrf.js';
+import email from './email/index.js';
+import financials from './financials/index.js';
+import locations from './locations/index.js';
+import login from './login/index.js';
+import logout from './logout.js';
+import signup from './signup/index.js';
 
 
 let router = express.Router();
