@@ -5,9 +5,9 @@ import httpErrors from 'http-errors';
 import moment from 'moment';
 import mongoose from 'mongoose';
 
-import csrf from '../../middleware/csrf';
-import multipartParser from '../../middleware/multipart-parser';
-import uuid from '../../util/uuid';
+import csrf from '../../middleware/csrf.js';
+import multipartParser from '../../middleware/multipart-parser.js';
+import uuid from '../../util/uuid.js';
 
 
 const AWSAccessKeyId = config.aws.credentials.AWSAccessKeyId;
