@@ -1,12 +1,12 @@
 'use strict';
 
+import config from 'config';
 import express from 'express';
 import moment from 'moment';
 
-import { Create as CreateSession } from '../../sessions';
+import { Create as CreateSession } from '../../sessions.js';
 import { LoginSessionTC } from '../../../schema/models/login-sessions.js';
 import { UserTC } from '../../../schema/models/users.js';
-import config from 'config';
 
 let router = express.Router();
 let login = router.route('/');
