@@ -5,7 +5,7 @@ import express from 'express';
 import moment from 'moment';
 
 import EmailClient from '../../extensions/mandrill/email-client.js';
-import uuid from '../../../lib/util/uuid';
+import uuid from '../../../lib/util/uuid.js';
 import { LoginSessionTC } from '../../../schema/models/login-sessions.js';
 import { UserTC } from '../../../schema/models/users.js';
 
