@@ -3,7 +3,7 @@
 import config from 'config';
 import express from 'express';
 
-import { Remove } from '../sessions';
+import { Remove } from '../sessions.js';
 
 let router = express.Router();
 let logout = router.route('/');
