@@ -1,14 +1,14 @@
 'use strict';
 
+import config from 'config';
 import express from 'express';
 import moment from 'moment';
 
-import uuid from '../../../lib/util/uuid';
-import { Create as CreateSession } from '../../sessions';
+import uuid from '../../../lib/util/uuid.js';
+import { Create as CreateSession } from '../../sessions.js';
 import { LoginSessionTC } from '../../../schema/models/login-sessions.js';
 import { UserTC } from '../../../schema/models/users.js';
-import config from 'config';
-import { PeopleTC } from '../../../schema/models/people';
+import { PeopleTC } from '../../../schema/models/people.js';
 
 let router = express.Router();
 let signup = router.route('/');
