@@ -1,10 +1,12 @@
 /* @flow */
 
 import _ from 'lodash';
-import composeWithMongoose from 'graphql-compose-mongoose/node8';
+import graphqlComposeMongoose from 'graphql-compose-mongoose';
 import mongoose from 'mongoose';
 
-import uuid from '../../lib/util/uuid';
+import uuid from '../../lib/util/uuid.js';
+
+const { composeWithMongoose } = graphqlComposeMongoose;
 
 
 export const EmailUpdateRequestSchema = new mongoose.Schema(
