@@ -3,12 +3,14 @@
 // TODO: FIXXX
 // TODO: geolocation type [Double]
 import _ from 'lodash';
-import composeWithMongoose from 'graphql-compose-mongoose/node8';
+import graphqlComposeMongoose from 'graphql-compose-mongoose';
 import httpErrors from 'http-errors';
 import moment from 'moment';
 import mongoose from 'mongoose';
 
-import uuid from "../../lib/util/uuid";
+import uuid from '../../lib/util/uuid.js';
+
+const { composeWithMongoose } = graphqlComposeMongoose;
 
 export const LocationsSchema = new mongoose.Schema(
 	{
