@@ -1,12 +1,15 @@
 /* @flow */
 
 import _ from 'lodash';
-import composeWithMongoose from 'graphql-compose-mongoose/node8';
+import graphqlComposeMongoose from 'graphql-compose-mongoose';
 import mongoose from 'mongoose'
 
-import uuid from '../../lib/util/uuid';
+import uuid from '../../lib/util/uuid.js';
 
-import { UserTC } from './users';
+import { UserTC } from './users.js';
+
+const { composeWithMongoose } = graphqlComposeMongoose;
+
 
 export const SessionSchema = new mongoose.Schema(
 	{
